backend/requirements.txt
fastapi
uvicorn[standard]
azure-identity
azure-cosmos
azure-storage-blob
azure-search-documents
openai
python-multipart
python-jose[cryptography]
python-docx
PyPDF2
pydantic
pydantic-settings
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp
azure-monitor-opentelemetry-exporter
pytest
pytest-asyncio
pytest-cov
ruff
